[mugen] copy time attribute in frame

+++ util/sdl/bitmap.cpp
@@ -98,9 +98,9 @@ static inline unsigned int alphaBlender(unsigned int x, unsigned int y, unsigned
     Uint8 blueY = 0;
     SDL_GetRGB(y, &format565, &redY, &greenY, &blueY);
 
-    int r = ((int) redY * dest + (int) redX * source) / 256;
-    int g = ((int) greenY * dest + (int) greenX * source) / 256;
-    int b = ((int) blueY * dest + (int) blueX * source) / 256;
+    int r = (redY * dest + redX * source) / 256;
+    int g = (greenY * dest + greenX * source) / 256;
+    int b = (blueY * dest + blueX * source) / 256;
 
     r = Util::min(r, 255);
     g = Util::min(g, 255);
